Allow SSLContext, disabling cert verificiation on __init__
Remove the method "ssl_no_cert_verify" from HttpClient,
instead allow setting the SSLContext or disabling the SSL certificate
verification on instantiation.
I prefer to provide an API with less state mutation of objects,
keep most customization to __init__.

If users would like to mutate state in the middle, it's still possible
to mutate the SSLContext, but I don't want to encourage that.